+++ src/com/chouguting/mipsemulator/hardware/Memory.java
@@ -6,10 +6,10 @@ import java.util.Map;
 public class Memory {
     //int[] data=new int[1073741824]; JAVA陣列不能宣告太大
     //用MAP來模擬Memory
-    Map<Integer, Integer> map = new HashMap<Integer, Integer>();
+    Map<Long, Long> map = new HashMap<Long, Long>();
 
-    public int getData(int index) {
-        Integer value = map.get(index);
+    public long getData(long index) {
+        Long value = map.get(index);
         if (value == null) {
             return 0;
         } else {
@@ -17,7 +17,7 @@ public class Memory {
         }
     }
 
-    public void setData(int index, int value) {
+    public void setData(long index, long value) {
         map.put(index, value);
     }
 }
